mammogram-mass-analyzer-v0.0/requirements.txt

# $ pip install -r requirements.txt

Flask==2.2.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
numpy==1.22.3
pandas==1.5.0
opencv-python==4.6.0.66
Pillow==9.3.0
PyYAML==6.0
pydicom==2.3.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.41.0
matplotlib==3.2.2
seaborn==0.11.0
scipy==1.4.1
